Reorganize tools
1) The readme is out-of-date, and unreadable on github
2) Some tools have their own directories, some don't
3) Many tools have neither readme nor descriptions.
4) Some files are misplaced

This reorganizes so that each tool has its own directory, even if it only has a single file
(Hint: If you use a tool, please add/update READMEs)

The master README is complete, and readable on github

The tools are in alphabetical order within category.  There are some cases where this probably isn't the right thing to do, e.g. where there are separate tools that do "to" and "from" conversions.

Each tool has at least a 1-line description in the master readme

This commit does not change any tool.