Fix fsio compilation on FreeBSD
Just tried to build fsio on FreeBSD and found it complaining about
unknown functions letoh32.  Turns out that the attempt to make this
portable to FreeBSD is incorrectly assuming that FreeBSD should be
like NetBSD, when in fact it is more like OpenBSD in this regard.

Moving the defined(__FreeBSD__) so it selects the same block as
__OpenBSD__ makes fsio build correctly on FreeBSD 11 (and presumably
later).